add ratelimit spam protection

+++ src/lavalink/index.js
@@ -2,7 +2,7 @@ const { Manager } = require('erela.js');
 const spotify = require('erela.js-spotify');
 const deezer = require('erela.js-deezer');
 const chalk = require('chalk');
-const { MessageEmbed } = require('discord.js');
+const { MessageEmbed, Collection } = require('discord.js');
 const {
     NODES,
     SPOTIFY_CLIENT_ID,
@@ -21,6 +21,8 @@ class lavalink extends Manager {
                 if (guild) guild.shard.send(payload)
             }
         })
+        client.rateLimit = new Collection();
+
         require('./player');
         this.on('nodeConnect', (node) => {
             console.log(chalk.green(`[LAVALINK] => [STATUS] ${node.options.identifier} successfully connected`))
@@ -63,7 +65,11 @@ class lavalink extends Manager {
             if (player.get('message') && !player.get('message').deleted) player.get('message').delete().catch(() => { });
             if (player.get('nowplaying')) { clearInterval(player.get('nowplaying')); player.get('nowplayingMSG').delete().catch(() => { }) }
         })
+        this.on('playerCreate', (player) => {
+            player.set('rateLimitStatus', { status: false })
+        })
         this.on('trackStart', (player, track) => {
+            if (player.get('rateLimitStatus').status == true) return;
             const channel = client.channels.cache.get(player.textChannel);
             const guild = client.guilds.cache.get(player.guild)
             var playembed = new MessageEmbed()
@@ -73,6 +79,7 @@ class lavalink extends Manager {
             channel.send(playembed).then(msg => player.set('message', msg))
         })
         this.on('trackEnd', (player, track) => {
+            if (player.get('rateLimitStatus').status == true) return;
             if (player.get('message') && !player.get('message').deleted) player.get('message').delete().catch(() => { });
             if (player.get('nowplaying')) { clearInterval(player.get('nowplaying')); player.get('nowplayingMSG').delete().catch(() => { }) }
         })
@@ -88,15 +95,38 @@ class lavalink extends Manager {
             if (player.get('nowplaying')) { clearInterval(player.get('nowplaying')); player.get('nowplayingMSG').delete().catch(() => { }) }
         })
         this.on('trackError', (player, track, payload) => {
-            if (player.get('error') && !player.get('error').deleted) player.get('error').delete().catch(() => { });
-            const channel = client.channels.cache.get(player.textChannel);
-            const guild = client.guilds.cache.get(player.guild)
-            var playembed = new MessageEmbed()
-                .setAuthor("Error")
-                .setDescription(`There was an error while playing **${track.title}** \n\`\`\`${payload.exception ? `Severity: ${payload.exception.severity}\nMessage: ${payload.exception.message}\nCause: ${payload.exception.cause}` : ''}\`\`\``)
-                .setColor(guild.me.displayHexColor != '#000000' ? guild.me.displayHexColor : '#00C7FF')
-            channel.send(playembed).then(msg => player.set('error', msg))
-            if (player.get('nowplaying')) { clearInterval(player.get('nowplaying')); player.get('nowplayingMSG').delete().catch(() => { }) }
+            const rate = client.rateLimit.get(player.guild)
+            const time1 = new Date()
+            const time2 = new Date()
+            if (rate && (time2 - rate.time <= 1000) && player.get('rateLimitStatus').status == false) {
+                const channel = client.channels.cache.get(player.textChannel);
+                const guild = client.guilds.cache.get(player.guild)
+                var playembed = new MessageEmbed()
+                    .setAuthor("Error")
+                    .setDescription(`Got a lot of errors within a short time. Now playing embed will be stopped for 40s to prevent spamming.`)
+                    .setColor(guild.me.displayHexColor != '#000000' ? guild.me.displayHexColor : '#00C7FF')
+                player.set('rateLimitStatus', { status: true })
+                setTimeout(() => {
+                    player.set('rateLimitStatus', { status: false })
+                }, 40000);
+                channel.send(playembed).then(msg => player.set('rateLimitMsg', msg))
+            } else if (player.get('rateLimitStatus').status == true) {
+                return
+            } else {
+                if (player.get('error') && !player.get('error').deleted) player.get('error').delete().catch(() => { });
+                if (player.get('rateLimitMsg') && !player.get('rateLimitMsg').deleted) player.get('rateLimitMsg').delete().catch(() => { });
+                const channel = client.channels.cache.get(player.textChannel);
+                const guild = client.guilds.cache.get(player.guild)
+                const err = payload.exception ? `Severity: ${payload.exception.severity}\nMessage: ${payload.exception.message}\nCause: ${payload.exception.cause}` : ''
+                var playembed = new MessageEmbed()
+                    .setAuthor("Error")
+                    .setDescription(`There was an error while playing **${track.title}** \n\`\`\`${err ? err : 'No error was provided from host'}\`\`\``)
+                    .setColor(guild.me.displayHexColor != '#000000' ? guild.me.displayHexColor : '#00C7FF')
+                channel.send(playembed).then(msg => player.set('error', msg))
+                if (player.get('nowplaying')) { clearInterval(player.get('nowplaying')); player.get('nowplayingMSG').delete().catch(() => { }) }
+            }
+            client.rateLimit.delete(player.guild)
+            client.rateLimit.set(player.guild, { time: time1 })
         })
         this.on('queueEnd', (player) => {
             const channel = client.channels.cache.get(player.textChannel);
